+++ modules/codegen/test/src/smithy4s/codegen/internals/SmithyBuildSpec.scala
@@ -17,14 +17,20 @@
 package smithy4s.codegen.internals
 
 import smithy4s.codegen.SmithyBuildJson
+import software.amazon.smithy.model.shapes.ShapeId
+import software.amazon.smithy.openapi.OpenApiVersion
+import scala.collection.immutable.ListSet
 
 final class SmithyBuildSpec extends munit.FunSuite {
   test("generate json") {
     val actual = SmithyBuild.writeJson(
-      SmithyBuild(
+      SmithyBuild.Serializable(
         "1.0",
-        List("src/"),
-        SmithyBuildMaven(List("dep"), List(SmithyBuildMavenRepository("repo")))
+        ListSet("src/"),
+        SmithyBuildMaven(
+          ListSet("dep"),
+          ListSet(SmithyBuildMavenRepository("repo"))
+        )
       )
     )
     assertEquals(
@@ -132,4 +138,66 @@ final class SmithyBuildSpec extends munit.FunSuite {
     )
   }
 
+  test("Can parse smithy-build file with OpenAPI plugin") {
+    val input =
+      """
+        |{
+        |   "version": "1.0",
+        |   "imports": [ "foo.smithy", "some/directory" ],
+        |   "plugins": {
+        |        "openapi": {
+        |            "service": "example.weather#Weather",
+        |            "version": "3.1.0"
+        |        }
+        |   }
+        |}""".stripMargin
+
+    val actual = io.circe.parser
+      .decode[SmithyBuild](input)
+      .left
+      .map(x => throw x)
+      .merge
+
+    // OpenApiConfig doesn't override equals, so we need to check expected == actual in pieces:
+    assertEquals(actual.maven, None)
+    assertEquals("1.0", actual.version)
+    assertEquals(
+      Set[os.FilePath](
+        os.FilePath("foo.smithy"),
+        os.FilePath("some/directory")
+      ),
+      actual.imports.toSet
+    )
+    val actualOpenApiConfig = actual
+      .getPlugin[SmithyBuildPlugin.OpenApi]
+      .getOrElse(fail("No OpenAPI plugin on parsed smithy-build"))
+      .config
+    assertEquals(
+      ShapeId.from("example.weather#Weather"),
+      actualOpenApiConfig.getService.toShapeId
+    )
+    assertEquals(
+      OpenApiVersion.VERSION_3_1_0,
+      actualOpenApiConfig.getVersion
+    )
+  }
+
+  test("Can parse smithy-build file with no optional fields") {
+    val input =
+      """
+        |{
+        |   "version": "1.0"
+        |}""".stripMargin
+
+    val actual = io.circe.parser
+      .decode[SmithyBuild](input)
+      .left
+      .map(x => throw x)
+      .merge
+
+    assertEquals(actual.maven, None)
+    assertEquals("1.0", actual.version)
+    assertEquals(Set.empty[os.FilePath], actual.imports.toSet)
+    assertEquals(Set.empty[SmithyBuildPlugin], actual.plugins.toSet)
+  }
 }