+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 * Constraints applied to list or map members are now correctly rendered in the generated code.
 * Makes the json decoding of tagged-unions lenient for AWS
+* Fix an issue with duplicated entries in generated smithy-build.json file (#1491)
+* Add support for passing custom OpenAPI config via a `smithy-build.json` file
+* Fix a bug when using `adt` with mixins, see #1457
 
 # 0.18.16